ai-engine/requirements.txt
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-core>=0.2.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pymongo>=4.6.0
